requirements.txt
flask
gunicorn
requests==2.22.0
cryptography
flask-socketio
eventlet==0.30.2
Flask-Failsafe
mysql-connector-python
Pillow
numpy

